resolve issue #38 (#45)
* simplify DiscoveryAllBuckets: remove a suspicious if
* fix warning from linter v1.60.3

+++ api.go
@@ -361,7 +361,7 @@ func (r *Router) RouterMapCallRWImpl(
 		return nil, err
 	}
 
-	if totalBucketCount != int32(r.cfg.TotalBucketCount) {
+	if uint64(totalBucketCount) != r.cfg.TotalBucketCount {
 		return nil, fmt.Errorf("unknown bucket counts %d", totalBucketCount)
 	}
